+++ SRC/DEVELOPMENT_RELEASE/MASTER/360/SRC/_GameVariables.c
@@ -12,7 +12,7 @@ WORD Player2Score=0;
 
 WORD GameState;			//For use by game
 
-unsigned long GameCount=0;		//Total number of games played
+DWORD GameCount=0;		//Total number of games played
 
 GameSettings CurrentGameSettings;
